feat(acid): rewrite secret handling
In the past, secret handling worked like this:

- Acid loads the Secret
- Acid runs the acid.js, but does not give it access to the secrets
- Acid.js must mount the Secret to the Pods it starts, and expose vars
  that way
- Pods have access to the Secret that has the project config in it

Since pods are less trusted than acid.js, it makes sense to reverse
this. So the present PR changes the workflow to work like this:

- Acid loads the Secret
- Acid sends ANYTHING IN THE `env` PART OF THE SECRET into acid.js
- acid.js then must send those into the Job's env field, and only those
  vals are sent to the Pod

In this model, pods NEVER have access to the project secret, and the
acid.js only has access to the contents of the secret's `env` object.

This provides the following features:

- An acid build can be run in a namespace other than the one Acid is
  running in.
- Acid.js never has access to secrets that the user doesn't grant it
  access to via the acid-project's `env` object

But as a side effect...

- The `Job.secrets` object is now obsolete and has been removed
- Anything in `Job.env` is added to the `pod.spec.container.env` in
  cleartext.

Closes #53