feat(worker): add error/after events
This commit refactors the what was in index.ts into two files:
index.ts and app.ts. The index.ts how has the minimal code to bootstrap
from the environment, and then it calls `app.App.run()`.

The app.ts now encapsulates the main runtime so, and is testable. New
tests have been added.

All of this was done in order to add two new events:

- 'after': Fires after run() has successfully executed an event.
- 'error': Fires if run() catches an uncaught or unresolved error.

These make it possible to register a global even to handle any error or
success.