ref(acid): rename 'acid' to 'acid-gateway'
This brings the webhook-providing service into better alignment with our
current naming scheme. It seems wrong that a non-critical piece of the
infrastructure was named 'acid', so I changed the name to indicate it's
role as a gateway to the core acid service.

This also makes it possible to rename `lsd` to `acid` if we want.