fix(worker): disable shared storage by default
Leaving this enabled adds a pretty serious performance hit on at least
some systems. And it is not typically used by containers out of the box.
That is, most of the time it is used, usage is already manually
configured by the script author. So it makes sense to go with
performance as the default metric.