fix(api,controller,worker): use ULID for ID
There were a few places where instead of using the ULID as an ID, the
system was using project + commit as an ID. That, however, is not
unique. This was resulting in incorrect data being returned by the API
server.

One of the reasons for the above was that job pods were not being
labeled with the ULID (or with the worker ID). So there was no way to
correlated a job with its worker parent or build. So this adds two new
labels on a job pod: worker and build. The build ID is also added to the
PVC for build storage.

Finally, the API server has been fixed to use the build ID instead of
the commit for finding associated jobs.

Closes #121