fix(api): show secret names, but not values (#128)
The API server now shows the names of the secrets, but not
their values. This allows user interfaces to show which secrets are
present, but not be able to show their values.

The returned value for secrets is still a map, to accurately reflect the
structure of a secret. But the value is always "REDACTED".