fix(brig): add verbose output for waiting
One of the common Brig problems occurs when Brig creates an event, but
the controller does not start the worker. This adds a single line of
verbose output indicating (a) when Brig starts waiting for the event,
and (b) what pod it expects to start as a worker.

Partially addresses #368