fix(worker): allow jobs during error handler
The error handler was not waiting for other events to finish. This
removes the process error code handling to the "exit" event, and allows
the "error" event to schedule jobs and other things that will require a
next-tick. It also intentionally bypasses the "after" event once an
error code is set. This ensures that if "error" triggers, "after" won't.

Closes #364