Improve brig api performance as seen in #433 by introducing caching (#446)
* add naive asynchronous fan in for ListWithLatestBuild

* Add apicache package which holds a cached versions of kube api objects
The apicache was introduced to solve performance issues when invoking the brigade api,
especially the route '/v1/projects-build' doesn't scale.
The issue was first introduced in: #433
The apicache packages exposes a single interface:
type ApiCache interface {
	GetSecretsFilteredBy(labelSelectors map[string]string) []v1.Secret
	GetPodsFilteredBy(labelSelectors map[string]string) []v1.Pod
}
It's design makes it easy to use for other api routes too.
However currently it only comes with two methods to solve the biggest performance issues.

* Add implementation for GetPodsFilteredBy

* Add implementation for GetSecretsFilteredBy

* Remove the naive 'optimization'
As I found out using benchmarks the problem lies in the fact
that way too many kube-api requests get invoked during the request.
This branch therefore introduces a new solution to the problem:
Simply watch the kubernetes api, automatically populate a local cache
and read only from the cache while answering the request.

* Update the GetProjectBuilds implementation to make use of the cache
This commit replaces the usage of the kubernetes api during the request
and replaces all its invocation with the cache instead.
For ease of use I've refactored the findWorker method signature to directly
use a slice of pods:
findWorker(id string, pods []v1.Pod)

* Add comment to relate to issue: https://github.com/kubernetes/client-go/issues/352
It seems like the implementation of fake.NewSimpleClientset() does not behave as expected.
Therefore in the current state the tests related to the cached api store won't pass.
This is because I did not yet found a way to mock objects into the watchers.

* Add apiCache object to the store to make it available from the api

* [WIP] trying to fix unit testing issues, might be obsolete

* Add package merge
This package offers utility methods to merge objects, e.g. channels

* Cleanup apicache/cache.go for readability
Structure code mode readable + add some comments.

* Add missing tests for apicache/cache.go to reach 100% coverage

* Add factory to apicache/pods.go

* Extract matching logic of string maps into listStore for reusability

* Add tests to apicache/pods_test.go to reach 100% coverage

* Add factory method to apicache/secrets.go

* Extract string maps matching logic to liststore.go for reusability

* Add tests to reach 100% coverage

* Add method to the store which blocks until all caches are synced
This is useful e.g. for testing when you'd like to get signaled
when the cache is populated.

* Extend the Store interface to expose the newly added BlockUntilApiCacheSynced
This method allows the user to wait until all caches are synced,
which is useful e.g. for testing.

* Add test invocation of s.BlockUntilApiCacheSynced

* Remove obsolete code from the tests

* Add BlockUntilApiCacheSynced to the mock store

* Add test invocation of the mocked BlockUntilApiCacheSynced func

* Add liststore.go which provides general purpose methods for ListStores
The package offers utiliy methods to create new list stores.
Aside from that it offers the stringMapsMatch method to match labels to runtime.Objects,
which is currently used in both the secret- and the pods-store.

* Add tests to reach 100% coverage for the liststore func's

* Add proper formatting and commenting

* Remove unused code