Eliminates a yarn race condition when installing brigade.json deps
`yarn` was previously invoked using via `child-process-promise` without
waiting for the promises to resolve, which led to a race between
multiple `yarn` instances attempting to update package.json, yarn.lock
and node_modules simultaneously.  This led to failed package
installation for brigade.json files with more than one dependency.

Now, `yarn` is invoked synchronously and in a single invocation, rather
than via multiple asynchronous execs.  This prevents the race condition
and makes package installation faster.

---

For testing, isolates all side-effects of prestart.js in an
`if (require.main === module block)`.

Uses Sinon and mock-require to get coverage of dependency discovery and
installation.  These tests are brittle, but since only the known-used
functions are mocked, if the implementation is changed, the tests should
begin to fail.  This is not as nice as a functional test that verifies
the desired outcome, but it's an acceptable compromise until that's
available.

Signed-off-by: Chris Chambers <chris.chambers@peanutcode.com>