Update QuickStart guide for v2
This keeps the quickstart simple. We get installed, define a project,
and trigger an event.
There is some content commented out currently that I am going to split
into a separate page that explains gateways and gets you set up.

One big change in the quickstart is that we don't link you off to
another page, for example when installing brigade. QuickStarts lose
people (i.e. they drop off and stop, or get lost) when they navigate to
another page and then have to go back and resume where they left off.
The goal for the quickstart (not all content) is to have a single page
that covers everything they need to know.

The installation is also specific to a dev cluster, whereas the
installation page will focus on a suitable production cluster.

I have also added corresponding instructions in powershell, so that
regardless of OS, people can follow along.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>